requirements.txt
certifi==2019.11.28
chardet==3.0.4
dnspython==1.16.0
future==0.18.2
idna==2.9
ijson==2.6.1
mysql-connector-python==8.0.19
oauthlib==3.1.0
protobuf==3.6.1
PyQt5==5.14.1
PyQt5-sip==12.7.1
python-twitter==3.5
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tabulate==0.8.7
urllib3==1.25.8
